WID-64: move functions that generate code to constants.ts
